Website completed
Left side done to amke the image and font accurate (image was zoomed in
initially but background-size: cover; fixed it).

right side was done made with form to create the log in we needed.

main.js created to include a passwords do not match for whenever a user
tries to input a password.